@@ -7,10 +7,10 @@
     "require": {
         "php": "^7.2|^8.0",
         "fideloper/proxy": "~4.0",
-        "guzzlehttp/guzzle": "^6.3",
+        "guzzlehttp/guzzle": "^7.0.1",
         "intervention/image": "^2.4",
         "laracasts/flash": "^3.0",
-        "laravel/framework": "^7.0",
+        "laravel/framework": "^8.0",
         "laravel/tinker": "^2.0",
         "predis/predis": "^2.1",
         "pusher/pusher-php-server": "^4.1",
@@ -20,7 +20,7 @@
         "filp/whoops": "~2.0",
         "fakerphp/faker": "^1.9.1",
         "mockery/mockery": "~1.0",
-        "phpunit/phpunit": "^8.5",
+        "phpunit/phpunit": "^9.0",
         "symfony/thanks": "^1.0"
     },
     "autoload": {
