@@ -1,20 +1,14 @@
 <?php
 
-use Illuminate\Http\Request;
-
+use App\Http\Controllers\ApiController;
 
 Route::group(['middleware' => 'api'], function () {
-	Route::post('char/items', 'ApiController@getItems');
-	Route::post('char/stats', 'ApiController@getStats');
-	Route::get('favorites/chars', 'ApiController@getFavsChars');
-	Route::get('twitch', 'ApiController@getTwitchChars');
-	Route::post('pob_code', 'ApiController@getXML');
-	Route::post('/build/save', 'ApiController@saveBuild');
-	Route::get('/build/{hash}', 'ApiController@getBuild');
-	Route::get('/snapshots/{acc}/{char}', 'ApiController@getSnapshots');
-
-	// Route::get('/getBuild', function() {
-	// 	$build = \App\Build::find(3);
-	// 	return $build;
-	// });
+	Route::post('char/items', [ApiController::class, 'getItems']);
+	Route::post('char/stats', [ApiController::class, 'getStats']);
+	Route::get('favorites/chars', [ApiController::class, 'getFavsChars']);
+	Route::get('twitch', [ApiController::class, 'getTwitchChars']);
+	Route::get('pob_code', [ApiController::class, 'getXML']);
+	Route::post('/build/save', [ApiController::class, 'saveBuild']);
+	Route::get('/build/{hash}', [ApiController::class, 'getBuild']);
+	Route::get('/snapshots/{acc}/{char}', [ApiController::class, 'getSnapshots']);
 });
